requirements.txt
Django~=3.1.1
Pillow
PyJWT
asgiref
certifi
chardet
django-crispy-forms
djangorestframework~=3.11.1
djangorestframework-simplejwt
essential-generators
idna
pip
setuptools
sqlparse
urllib3
requests~=2.24.0
names~=0.3.0
toml~=0.10.1